+++ js/data.js
@@ -1,0 +1,49 @@
+import {getRandomInteger} from './util.js';
+
+const MIN_AVATAR_NAMBER = 1;
+const MAX_AVATAR_NAMBER = 6;
+const MIN_LIKES_NAMBER = 15;
+const MAX_LIKES_NAMBER = 200;
+const OBJECTS_COUNT = 25;
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+const MIN_MESSAGES_NUMBER = 0;
+const MAX_MESSAGE_NUMBER = MESSAGES.length - 1;
+const NAMES = [
+  'Иван',
+  'Хуан Себастьян',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+  'Люпита',
+  'Вашингтон',
+];
+const MIN_NAMES_NUMBER = 0;
+const MAX_NAMES_NUMBER = NAMES.length - 1;
+
+const createCommentGenerator = (id) => (
+  {
+    id: id,
+    avatar: `img/avatar-${getRandomInteger(MIN_AVATAR_NAMBER, MAX_AVATAR_NAMBER)}.svg`,
+    message: MESSAGES[getRandomInteger(MIN_MESSAGES_NUMBER, MAX_MESSAGE_NUMBER)],
+    name: NAMES[getRandomInteger(MIN_NAMES_NUMBER, MAX_NAMES_NUMBER)],
+  });
+
+const createIdGenerator = (id) => ({
+  id: id,
+  url: `photos/${id}.jpg`,
+  description: 'Лето, день чудесный',
+  likes: getRandomInteger(MIN_LIKES_NAMBER, MAX_LIKES_NAMBER),
+  comments: [createCommentGenerator(id)],
+});
+
+const createImages = () => Array.from({length: OBJECTS_COUNT}, (v, k) => createIdGenerator(k + 1));
+
+export {createImages};
